# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Dodo Invoice & Payment Service
  version: 0.1.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: invalid_state
            message:
              type: string
            details:
              type: object
              nullable: true
...
    LineItem:
      type: object
      required: [description, quantity, unit_amount_cents]
      properties:
        description: { type: string }
        quantity: { type: integer, minimum: 1 }
        unit_amount_cents: { type: integer, minimum: 0 }
    Invoice:
      type: object
      properties:
        id: { type: string, format: uuid }
        customer_id: { type: string, format: uuid }
        state:
          type: string
          enum: [draft, open, paid, void, uncollectible]
        total_cents: { type: integer }
        due_date: { type: string, format: date }
        line_items:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/LineItem'
              - type: object
                properties:
                  line_total_cents: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PayResponse:
      type: object
      properties:
        payment_attempt:
          type: object
          properties:
            id: { type: string, format: uuid }
            invoice_id: { type: string, format: uuid }
            state: { type: string, enum: [pending, succeeded, failed] }
            psp_ref: { type: string, nullable: true }
            failure_code: { type: string, nullable: true }
        invoice:
          $ref: '#/components/schemas/Invoice'
paths:
...
  /invoices/{id}/pay:
    post:
      summary: Pay invoice
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [card_token]
              properties:
                card_token: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayResponse'
        '202':
          description: Payment pending (PSP timeout)
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...